backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Seu Nome <email@exemplo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "^6.0"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
python-decouple = "^3.8"
faker = "^38.2.0"

[dependency-groups]
docs = [
    "mkdocs (>=1.6.1,<2.0.0)"
]
